refactor: use release notes earliest and latest dates for from and to date rendering
